refactor: [FE] NetworkServiceUtils 파일명 변경
- Intent
소켓 코드는 전부 따로 분리하고 axios 관련 코드만 남게 되어 변경했어요.
