+++ frontend/src/axios/sharedResult.ts
